Add interlocked pallet layout support
