+++ packing_app/core/algorithms.py
@@ -67,6 +67,34 @@ def pack_rectangles_mixed_greedy(width, height, wprod, lprod, margin=0):
             best_positions = temp_positions
     return best_count, best_positions
 
+def compute_interlocked_layout(pallet_w, pallet_l, box_w, box_l, num_layers=4):
+    """Return positions for standard and interlocked stacking."""
+    count, base_positions = pack_rectangles_mixed_greedy(pallet_w, pallet_l, box_w, box_l)
+
+    base_layers = [base_positions for _ in range(num_layers)]
+
+    min_x = min(x for x, y, w, h in base_positions)
+    max_x = max(x + w for x, y, w, h in base_positions)
+    min_y = min(y for x, y, w, h in base_positions)
+    max_y = max(y + h for x, y, w, h in base_positions)
+
+    shift_x = 0.0
+    shift_y = 0.0
+    if min_x >= box_w / 2 and max_x + box_w / 2 <= pallet_w:
+        shift_x = box_w / 2
+    elif min_y >= box_l / 2 and max_y + box_l / 2 <= pallet_l:
+        shift_y = box_l / 2
+
+    interlocked_layers = []
+    for layer_idx in range(num_layers):
+        if layer_idx % 2 == 0:
+            interlocked_layers.append(base_positions)
+        else:
+            shifted = [(x + shift_x, y + shift_y, w, h) for x, y, w, h in base_positions]
+            interlocked_layers.append(shifted)
+
+    return count, base_layers, interlocked_layers
+
 def pack_rectangles_mixed_max(width, height, wprod, lprod, margin=0):
     eff_width = width - margin
     eff_height = height - margin
