+++ packing_app/gui/tab_pallet.py
@@ -4,7 +4,7 @@ import matplotlib
 matplotlib.use("TkAgg")
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
-from packing_app.core.algorithms import pack_rectangles_mixed_greedy
+from packing_app.core.algorithms import pack_rectangles_mixed_greedy, compute_interlocked_layout
 from core.utils import load_cartons, load_pallets
 
 
@@ -341,6 +341,10 @@ class TabPallet(ttk.Frame):
         count2, positions2 = pack_rectangles_mixed_greedy(pallet_w, pallet_l, box_l, box_w)
         positions2 = self.center_layout(positions2, pallet_w, pallet_l)
         self.layouts.append((count2, positions2, "Naprzemienny"))
+
+        _, _, interlocked_layers = compute_interlocked_layout(pallet_w, pallet_l, box_w, box_l, num_layers=2)
+        shifted = self.center_layout(interlocked_layers[1], pallet_w, pallet_l)
+        self.layouts.append((len(shifted), shifted, "Przesunięty"))
 
         self.layout_map = {name: idx for idx, (_, __, name) in enumerate(self.layouts)}
         self.update_transform_frame()
